refactor: community 컴포넌트 분리 및 메모이제이션
