+++ app/community/_component/BottomBar.tsx
@@ -1,18 +1,14 @@
+import {memo} from 'react';
 import Image from 'next/image';
 
-export default function BottomBar({style}: {style?: string}) {
+const BottomBar = memo(function BottomBar() {
   return (
-    <div
-      className={`${style} relative overflow-hidden w-[111vw] h-[10vw] left-[-5vw]  flex justify-center`}
-    >
-      <div className="absolute left-0 top-0 w-[111vw] aspect-[2132/585] z-10">
-        <Image
-          fill
-          src="/images/community/bottom.webp"
-          priority
-          alt="bottom bar"
-        />
+    <div className="relative overflow-hidden w-[111vw]  h-[10vw] left-[-5vw] top-[-9vw] flex justify-center">
+      <div className="absolute left-0 top-0 w-[111vw] aspect-[2132/585] z-0">
+        <Image fill src="/images/community/bottom.webp" alt="" />
       </div>
     </div>
   );
-}
+});
+
+export default BottomBar;
