+++ app/community/_component/Info.tsx
@@ -8,14 +8,15 @@ import RankBox from '@/components/ranking/RankBox';
 import Slider from '@/components/ranking/Slider';
 import {rankType, universityType} from '@/type';
 import Image from 'next/image';
-import {useRouter} from 'next/navigation';
 import {useEffect, useState} from 'react';
+import {Bg1, Bg2} from './Bg';
+import SearchArea from './SearchArea';
+import BottomBar from './BottomBar';
 
 export default function Info() {
   const [rankStandard, setRankStandard] = useState<rankType>('weekly');
   const [data, setData] = useState<universityType[]>([]);
   const {ref, isVisible} = useScrollAnimation();
-  const router = useRouter();
 
   useEffect(() => {
     getData();
@@ -70,52 +71,10 @@ export default function Info() {
     ]);
   };
 
-  const handleKeyDown = (e: React.KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter' && e.currentTarget.value) {
-      const keyword = e.currentTarget.value.trim();
-      router.push(
-        `/community/${encodeURIComponent(keyword.replace('학교', ''))}`,
-      );
-    }
-  };
-
   return (
     <div className="w-screen relative bg-black overflow-hidden">
-      <div className="relative w-screen aspect-[1954/1524]">
-        <Image
-          src="/images/main/bg_page5.webp"
-          fill
-          sizes="100vw"
-          alt="bg"
-          priority
-          fetchPriority="high"
-          className="absolute z-0 top-[-17vw]"
-        />
-      </div>
-
-      <div className="absolute w-screen h-screen top-[10vw]">
-        <div className="w-screen flex flex-col items-center justify-center absolute z-10">
-          <h1 className="text-white text-[1.875vw]">
-            대전과 챌린지로 학교 순위 올리기
-          </h1>
-          <p className="text-grey2 text-[1vw] mt-[0.6vw]">
-            현재 상위 학교들의 랭킹과 순위 지표를 확인하고 대전을 통해 순위를
-            올려보세요!
-          </p>
-
-          <div className="mt-[2.8vw] w-[43vw] aspect-[837/67] relative text-[1vw] text-white border border-communityGrey rounded-[3vw] flex justify-center items-center">
-            <input
-              className="w-full bg-inherit border-none text-center placeholder:text-white outline-none"
-              placeholder="학교 이름을 검색해보세요"
-              onKeyDown={handleKeyDown}
-            />
-            <div className="absolute w-[1.2vw] aspect-[1/1] right-[1vw]">
-              <Image fill src="/svg/search.svg" alt={'search'} />
-            </div>
-          </div>
-        </div>
-      </div>
-
+      <Bg1 />
+      <SearchArea />
       <div className="w-full flex justify-between items-center absolute top-[40vw] z-30">
         <div className="ml-[6.25vw] flex flex-col gap-[3.28vw]">
           <LeftBox1 isVisible={isVisible} />
@@ -133,23 +92,8 @@ export default function Info() {
 
       <Slider top={'top-[46vw]'} data={data} />
 
-      <div className="absolute top-[25vw] left-[6.5vw] z-10">
-        <div className="relative w-[87.5vw] aspect-[1680/830]">
-          <Image
-            src="/images/main/overlay_page5.webp"
-            priority
-            fetchPriority="high"
-            alt="bg"
-            fill
-          />
-        </div>
-      </div>
-
-      <div className="relative overflow-hidden w-[111vw]  h-[10vw] left-[-5vw] top-[-9vw] flex justify-center">
-        <div className="absolute left-0 top-0 w-[111vw] aspect-[2132/585] z-0">
-          <Image fill src="/images/community/bottom.webp" alt="" />
-        </div>
-      </div>
+      <Bg2 />
+      <BottomBar />
 
       <div ref={ref} className="w-full bg-white absolute bottom-[20vw]" />
     </div>
