+++ app/community/_component/UnivMapArea.tsx
@@ -1,9 +1,9 @@
 'use client';
 
-import Image from 'next/image';
 import UnivMap from './UnivMap';
 import {useEffect, useState} from 'react';
 import {communityUnivType} from '@/type';
+import Character from './Character';
 // import {getRank} from '@/app/apis/rankApi';
 
 export default function UnivMapArea() {
@@ -203,20 +203,7 @@ export default function UnivMapArea() {
     <div className="w-screen">
       <div className="relative w-full aspect-[1920/1020] overflow-hidden z-10">
         <UnivMap data={data} />
-        <div className="absolute top-[5vw] left-[5vw] flex gap-[0.6vw]">
-          <div className="relative w-[6vw] aspect-[1/1]">
-            <Image
-              src="/images/characters/wink.png"
-              alt="character"
-              fill
-              sizes="6vw"
-              className="object-contain w-[21vw] aspect-[402/406] transform scale-x-[-1]"
-            />
-          </div>
-          <div className="w-[20.5vw] h-[2.7vw] bg-dark1 text-white rounded-[0.468vw] text-[0.9vw] flex justify-center items-center">
-            랭킹으로 점수를 높여 학교 영역을 넓혀보세요
-          </div>
-        </div>
+        <Character />
       </div>
     </div>
   );
